requirements.txt
torch
torchvision
scikit-learn
opencv-python
pandas
matplotlib
seaborn
tqdm
joblib
scikit-image
Pillow